+++ todo/src/main/java/com/todo/todo/domain/task/service/TaskServiceImpl.java
@@ -1,13 +1,13 @@
-package com.todo.todo.domain.task.servise;
+package com.todo.todo.domain.task.service;
 
+import com.todo.todo.domain.project.repository.ProjectRepository;
 import com.todo.todo.domain.task.entity.Task;
-import com.todo.todo.infrastructure.exceptions.custom.TaskNotFoundException;
 import com.todo.todo.domain.task.mapper.TaskMapper;
 import com.todo.todo.domain.task.models.TaskCreateDTO;
 import com.todo.todo.domain.task.models.TaskResponseDTO;
 import com.todo.todo.domain.task.models.TaskUpdateDTO;
-import com.todo.todo.domain.project.repository.ProjectRepository;
 import com.todo.todo.domain.task.repository.TaskRepository;
+import com.todo.todo.infrastructure.exceptions.custom.TaskNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -22,13 +22,11 @@ public class TaskServiceImpl implements TaskService {
     @Override
     public TaskResponseDTO save(TaskCreateDTO task) {
         Task taskForSave = taskMapper.map(task);
+
         taskForSave.setProject(projectRepository.findById(task.projectId()).get());
         Task newTask = taskRepo.save(taskForSave);
         return taskMapper.map(newTask);
     }
-
-
-
     @Autowired
     public void setTaskRepo(TaskRepository taskRepo) {
         this.taskRepo = taskRepo;
@@ -39,8 +37,7 @@ public class TaskServiceImpl implements TaskService {
         return taskMapper.map(taskRepo.findById(id).orElseThrow(() -> new TaskNotFoundException(id)));
     }
 
-
-   @Override
+    @Override
     public TaskResponseDTO update(TaskUpdateDTO task) {
         Task dbTask = taskRepo.findById(task.id()).orElseThrow(() -> new TaskNotFoundException(task.id()));
         taskMapper.map(task, dbTask);
@@ -50,14 +47,17 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public void delete(Long id) {
-        taskRepo.deleteById(id);
+        if (taskRepo.existsById(id)) {
+            taskRepo.deleteById(id);
+        } else {
+            throw new TaskNotFoundException(id);
+        }
     }
 
     @Override
     public Page<TaskResponseDTO> getAll(Pageable pageable) {
         return taskRepo.findAll(pageable).map(taskMapper::map);
     }
-
 
     @Autowired
     public void setTaskMapper(TaskMapper taskMapper) {
